+++ src/main/java/com/seungki/urlshortener/api/controller/ApiExceptionHandler.java
@@ -1,8 +1,14 @@
-package com.seungki.urlshortener.api.controller.exception;
+package com.seungki.urlshortener.api.controller;
 
-import com.seungki.urlshortener.web.exception.ShortcodeGenerationException;
-import com.seungki.urlshortener.web.exception.ShortcodeNotFoundException;
-import com.seungki.urlshortener.web.exception.UrlNotFoundException;
+import static com.seungki.urlshortener.api.dto.ErrorCode.SHORTCODE_GENERATION_FAILED;
+import static com.seungki.urlshortener.api.dto.ErrorCode.SHORTCODE_NOT_FOUND;
+import static com.seungki.urlshortener.api.dto.ErrorCode.URL_NOT_FOUND;
+import static com.seungki.urlshortener.api.dto.ErrorCode.VALIDATION_ERROR;
+
+import com.seungki.urlshortener.api.dto.ApiErrorResponse;
+import com.seungki.urlshortener.common.exception.ShortcodeGenerationException;
+import com.seungki.urlshortener.common.exception.ShortcodeNotFoundException;
+import com.seungki.urlshortener.common.exception.UrlNotFoundException;
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -20,35 +26,43 @@ public class ApiExceptionHandler {
     @ExceptionHandler(ShortcodeNotFoundException.class)
     public ResponseEntity<ApiErrorResponse> handleShortcodeNotFoundException(ShortcodeNotFoundException ex) {
         log.info("[ShortcodeNotFoundException] message: {}", ex.getMessage());
-        ApiErrorResponse response = new ApiErrorResponse(ErrorCode.SHORTCODE_NOT_FOUND, ex.getMessage());
-        return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+        return new ResponseEntity<>(
+                new ApiErrorResponse(SHORTCODE_NOT_FOUND, ex.getMessage()),
+                HttpStatus.NOT_FOUND
+        );
     }
 
     @ExceptionHandler(UrlNotFoundException.class)
     public ResponseEntity<ApiErrorResponse> handleUrlNotFoundException(UrlNotFoundException ex) {
         log.info("[UrlNotFoundException] message: {}", ex.getMessage());
-        ApiErrorResponse response = new ApiErrorResponse(ErrorCode.URL_NOT_FOUND, ex.getMessage());
-        return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+        return new ResponseEntity<>(
+                new ApiErrorResponse(URL_NOT_FOUND, ex.getMessage()),
+                HttpStatus.NOT_FOUND
+        );
     }
 
     @ExceptionHandler(ShortcodeGenerationException.class)
     public ResponseEntity<ApiErrorResponse> handleShortcodeGenerationException(ShortcodeGenerationException ex) {
         log.info("[ShortcodeGenerationException] message: {}", ex.getMessage());
-        ApiErrorResponse response = new ApiErrorResponse(ErrorCode.SHORTCODE_GENERATION_FAILED, ex.getMessage());
-        return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+        return new ResponseEntity<>(
+                new ApiErrorResponse(SHORTCODE_GENERATION_FAILED, ex.getMessage()),
+                HttpStatus.INTERNAL_SERVER_ERROR
+        );
     }
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
     public ResponseEntity<ApiErrorResponse> handleValidationExceptions(MethodArgumentNotValidException ex) {
         log.info("[ValidationException] message: {}", ex.getMessage());
+
         List<String> errors = ex.getBindingResult()
                 .getAllErrors()
                 .stream()
                 .map(error -> ((FieldError) error).getField() + ": " + error.getDefaultMessage())
                 .toList();
-        ApiErrorResponse response = new ApiErrorResponse(ErrorCode.VALIDATION_ERROR, ex.getMessage());
-        response.addDetail("errors", errors);
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+
+        ApiErrorResponse errorResponse = new ApiErrorResponse(VALIDATION_ERROR, ex.getMessage());
+        errorResponse.addDetail("errors", errors);
+        return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
     }
 
 }
