+++ src/main/java/com/seungki/urlshortener/api/dto/ApiErrorResponse.java
@@ -1,4 +1,4 @@
-package com.seungki.urlshortener.api.controller.exception;
+package com.seungki.urlshortener.api.dto;
 
 import java.time.LocalDateTime;
 import java.util.HashMap;
