+++ src/main/java/com/seungki/urlshortener/api/dto/UrlShortenRequest.java
@@ -1,4 +1,4 @@
-package com.seungki.urlshortener.web.controller.dto;
+package com.seungki.urlshortener.api.dto;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Pattern;
@@ -6,7 +6,7 @@ import lombok.Getter;
 
 @Getter
 public class UrlShortenRequest {
-    @NotEmpty(message = "URL은 공백을 허용하지 않습니다")
+    @NotEmpty(message = "공백을 허용하지 않습니다")
     @Pattern.List({
             @Pattern(
                     regexp = "^(http://|https://).*",
@@ -14,7 +14,7 @@ public class UrlShortenRequest {
             ),
             @Pattern(
                     regexp = "[a-zA-Z0-9-._~:/?#@!$&'()*+,;=%]+",
-                    message = "URL은 영문자, 숫자 그리고 특수 문자(._~:/?#@!$&'()*+,;=%)의 조합만 허용합니다"
+                    message = "URL은 영문자, 숫자 그리고 특수 문자(._~:/?#@!$&'()*+,;=%)만 허용합니다"
             )
     })
     private String url;
