+++ src/main/java/com/seungki/urlshortener/api/dto/UrlShortenResponse.java
@@ -1,11 +1,11 @@
-package com.seungki.urlshortener.web.controller.dto;
+package com.seungki.urlshortener.api.dto;
 
-import com.seungki.urlshortener.web.domain.UrlMapping;
+import com.seungki.urlshortener.common.domain.UrlMapping;
 import java.time.LocalDateTime;
 import lombok.Getter;
 
 @Getter
-public class UrlDetailResponse {
+public class UrlShortenResponse {
 
     private String shortcode;
     private String originalUrl;
@@ -13,12 +13,11 @@ public class UrlDetailResponse {
     private LocalDateTime viewedAt;
     private int viewCount;
 
-    public UrlDetailResponse(UrlMapping urlMapping) {
+    public UrlShortenResponse(UrlMapping urlMapping) {
         shortcode = urlMapping.getShortcode();
         originalUrl = urlMapping.getOriginalUrl();
         createdAt = urlMapping.getCreatedAt();
         viewedAt = urlMapping.getViewedAt();
         viewCount = urlMapping.getViewCount();
     }
-
 }
