+++ src/main/java/com/seungki/urlshortener/common/exception/ShortcodeNotFoundException.java
@@ -1,4 +1,4 @@
-package com.seungki.urlshortener.web.exception;
+package com.seungki.urlshortener.common.exception;
 
 public class ShortcodeNotFoundException extends RuntimeException {
     public ShortcodeNotFoundException() {
