+++ src/main/java/com/seungki/urlshortener/web/dto/UrlDetailResponse.java
@@ -1,11 +1,11 @@
-package com.seungki.urlshortener.api.controller.dto;
+package com.seungki.urlshortener.web.dto;
 
-import com.seungki.urlshortener.web.domain.UrlMapping;
+import com.seungki.urlshortener.common.domain.UrlMapping;
 import java.time.LocalDateTime;
 import lombok.Getter;
 
 @Getter
-public class UrlShortenResponse {
+public class UrlDetailResponse {
 
     private String shortcode;
     private String originalUrl;
@@ -13,11 +13,12 @@ public class UrlShortenResponse {
     private LocalDateTime viewedAt;
     private int viewCount;
 
-    public UrlShortenResponse(UrlMapping urlMapping) {
+    public UrlDetailResponse(UrlMapping urlMapping) {
         shortcode = urlMapping.getShortcode();
         originalUrl = urlMapping.getOriginalUrl();
         createdAt = urlMapping.getCreatedAt();
         viewedAt = urlMapping.getViewedAt();
         viewCount = urlMapping.getViewCount();
     }
+
 }
