+++ src/main/resources/application.yml
@@ -4,13 +4,13 @@ spring:
     username: sa
     password:
     driver-class-name: org.h2.Driver
-
-  jpa:
+    
+jpa:
+  hibernate:
+    ddl-auto: create
+  properties:
     hibernate:
-      ddl-auto: create
-    properties:
-      hibernate:
-        format_sql: true
+      format_sql: true
 
 logging.level:
   org.hibernate.SQL: debug
